refactor(research): added hits to its own component searchhits
