+++ apps/research/src/components/search/index.ts
@@ -1,1 +1,2 @@
 export { default as SearchText } from './SearchText';
+export { default as SearchHits } from './SearchHits';
